pyproject.toml
[build-system]
requires = ["setuptools>=75.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "LLM-API-Endpoints"
version = "0.1.1"
authors = [
  { name = "Luiz Pedro Franciscatto Guerra" },
]
dependencies = [
    "python-dotenv==1.0.1",
    "openai==1.60.1",
    "google-generativeai==0.8.4",
    "google-api-core==2.24.0",
    "google-genai===1.2.0",
    "protobuf==5.29.3",
]